+++ src/main/java/roomescape/controller/api/ReservationApiController.java
@@ -2,9 +2,6 @@ package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
-import java.net.URI;
-import java.time.LocalDate;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,35 +9,32 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
-import roomescape.service.dto.request.ReservationAdminSaveRequest;
 import roomescape.service.dto.request.ReservationSaveRequest;
 import roomescape.service.dto.response.ReservationResponse;
 import roomescape.service.dto.response.UserReservationResponse;
-import roomescape.service.reservation.AdminReservationCreateService;
 import roomescape.service.reservation.ReservationCreateService;
 import roomescape.service.reservation.ReservationDeleteService;
 import roomescape.service.reservation.ReservationFindService;
+
+import java.net.URI;
+import java.util.List;
 
 @Validated
 @RestController
 public class ReservationApiController {
 
     private final ReservationCreateService reservationCreateService;
-    private final AdminReservationCreateService adminReservationCreateService;
     private final ReservationFindService reservationFindService;
     private final ReservationDeleteService reservationDeleteService;
 
     public ReservationApiController(ReservationCreateService reservationCreateService,
-                                    AdminReservationCreateService adminReservationCreateService,
                                     ReservationFindService reservationFindService,
                                     ReservationDeleteService reservationDeleteService) {
         this.reservationCreateService = reservationCreateService;
-        this.adminReservationCreateService = adminReservationCreateService;
         this.reservationFindService = reservationFindService;
         this.reservationDeleteService = reservationDeleteService;
     }
@@ -48,20 +42,6 @@ public class ReservationApiController {
     @GetMapping("/reservations")
     public ResponseEntity<List<ReservationResponse>> getReservations() {
         List<Reservation> reservations = reservationFindService.findReservations();
-        return ResponseEntity.ok(
-                reservations.stream()
-                        .map(ReservationResponse::new)
-                        .toList()
-        );
-    }
-
-    @GetMapping("/admin/reservations/search")
-    public ResponseEntity<List<ReservationResponse>> getSearchingReservations(@RequestParam long memberId,
-                                                                              @RequestParam long themeId,
-                                                                              @RequestParam LocalDate dateFrom,
-                                                                              @RequestParam LocalDate dateTo
-    ) {
-        List<Reservation> reservations = reservationFindService.searchReservations(memberId, themeId, dateFrom, dateTo);
         return ResponseEntity.ok(
                 reservations.stream()
                         .map(ReservationResponse::new)
@@ -84,13 +64,6 @@ public class ReservationApiController {
                                                                     @AuthenticatedMember Member member) {
         Reservation newReservation = reservationCreateService.createReservation(request, member);
         return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId()))
-                .body(new ReservationResponse(newReservation));
-    }
-
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> addReservationByAdmin(@RequestBody @Valid ReservationAdminSaveRequest request) {
-        Reservation newReservation = adminReservationCreateService.createReservation(request);
-        return ResponseEntity.created(URI.create("/admin/reservations/" + newReservation.getId()))
                 .body(new ReservationResponse(newReservation));
     }
